pyproject.toml
[project]
name = "pi-grpo"
version = "0.1.0"
description = "Physics-informed RL: PPO + DPO + GRPO with a physics-aware reward model"
authors = [{ name = "Arun Sharma", email = "arunshar@umn.edu" }]
requires-python = ">=3.11,<3.13"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "httpx>=0.27",
  "asyncpg>=0.29",
  "redis>=5.0",
  "torch>=2.4",
  "transformers>=4.44",
  "trl>=0.10",
  "accelerate>=0.34",
  "deepspeed>=0.15",
  "datasets>=2.20",
  "peft>=0.12",
  "bitsandbytes>=0.43",
  "sentencepiece>=0.2",
  "tiktoken>=0.7",
  "numpy>=1.26",
  "scipy>=1.13",
  "shapely>=2.0",
  "structlog>=24.1",
  "tenacity>=8.3",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-exporter-otlp>=1.25",
  "wandb>=0.17",
  "ray[default]>=2.34",
  "vllm>=0.5.3",
  "orjson>=3.10",
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-cov>=5.0",
  "ruff>=0.5",
  "mypy>=1.10",
  "hypothesis>=6.108",
]
frontend = ["streamlit>=1.36", "plotly>=5.22"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 120
target-version = "py311"
[tool.ruff.lint]
select = ["E","F","I","UP","B","SIM","RUF"]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E501"]
"app/prompts/templates.py" = ["RUF001"]
[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q --strict-markers"
[tool.coverage.run]
branch = true
source = ["app", "observability"]
omit = [
  "app/agents/tools/*",                       # tool stubs
  "app/trainers/*",                           # GPU-only; covered on the GPU CI lane
  "app/rollouts/*",                           # GPU-only
  "app/agents/trainer_agent.py",
  "app/services/checkpoint.py",               # no torch in the CPU test env
  "app/services/data_pipeline.py",
  "app/services/eval_runner.py",
  "app/services/token_optimizer.py",          # CPU-only stub
  "app/components/preference_builder.py",     # exercised by build_preferences.py CLI
  "app/components/pidpm_scorer.py",           # requires a Pi-DPM checkpoint
  "app/reward_models/preference_reward.py",   # nn.Module forward pass test moves to GPU lane
  "app/prompts/*",                            # static templates
  "app/security/*",                           # exercised via FastAPI tests; CPU is sufficient
]
[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
